#67: Moved the cmd out of pkgs
